pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmscout"
dynamic = ["version"]
description = 'LLM Proxy that automatically scans LLM endpoints from Cloud providers'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Cedric Vidal", email = "cedric@vidal.biz" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
  "python-dotenv==1.0.1",
  "PyYAML==6.0.1",
  "coloredlogs==15.0.1",
  "azureml-core",
  "msrestazure",
  "azure-mgmt-resource",
  "azure-identity==1.17.1",
  "pytest==8.1.2",
  "requests==2.32.3",
  "click==8.1.7",
  "rich==13.8.1",
  "rich-click==1.8.3",
  "survey==5.4.0",
  "litellm[proxy]==1.48.14",
]

[project.urls]
Documentation = "https://github.com/Cedric Vidal/llmscout#readme"
Issues = "https://github.com/Cedric Vidal/llmscout/issues"
Source = "https://github.com/Cedric Vidal/llmscout"

[project.scripts]
llmscout = "llmscout.cli:llmscout"

[tool.hatch.version]
path = "src/llmscout/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/llmscout tests}"

[tool.coverage.run]
source_pkgs = ["llmscout", "tests"]
branch = true
parallel = true
omit = [
  "src/llmscout/__about__.py",
]

[tool.coverage.paths]
llmscout = ["src/llmscout", "*/llmscout/src/llmscout"]
tests = ["tests", "*/llmscout/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
